+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   - `Button`, `Button'2`, `Button'3`...
   - `Context.Provider` → `AnonymousContext.Provider`, `AnonymousContext'2.Provider`...
 - Added output for received bytes of events data in status bar
+- Fixed an issue with solutions like `react-refresh` that might not work when using RRT without React Devtools
 
 ## 0.3.0 (September 8, 2021)